k-plane delete regularization

+++ plenoxels/raymarching/spatial_distortions.py
@@ -41,10 +41,10 @@ class SceneContraction(SpatialDistortion):
         super().__init__()
         self.order = order
         if global_translation is None:
-            global_translation = torch.tensor([0.0, 0.0, 0.0])
+            global_translation = torch.tensor([0.0, 0.0, 0.0],dtype=torch.float32)
         self.global_translation = nn.Parameter(global_translation, requires_grad=False)
         if global_scale is None:
-            global_scale = torch.tensor([1.0, 1.0, 1.0])
+            global_scale = torch.tensor([1.0, 1.0, 1.0],dtype=torch.float32)
             
         self.global_scale = nn.Parameter(global_scale, requires_grad=False)
